+++ substrace_lints/src/substrace_lints/mod.rs
@@ -1,4 +1,5 @@
 pub mod extrinsics_must_be_tagged;
 pub mod missing_security_doc;
+pub mod missing_transactional;
 pub mod no_panics;
 mod auxiliary;
